agents/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.0.50
pydantic>=2.0.0
temporalio>=1.5.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
nest-asyncio>=1.5.0
pytest>=7.0.0
pytest-asyncio>=0.21.0